refactor to reduce group file size
